learning-path/W03/experiments/requirements.txt
# Job Fitment Analysis Agent - Requirements
# Python 3.8+ required

# Core ML/AI
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
bert-score>=0.3.13

# Vector Database
faiss-cpu>=1.7.4

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# API & Authentication
huggingface_hub>=0.16.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web Scraping (for future job board integration)
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0

# Utilities
tqdm>=4.65.0

# Optional: For enhanced features
# langchain>=0.1.0
# chromadb>=0.4.0
# selenium>=4.15.0  # For dynamic job board scraping
